feat: Add Tracker class for object tracking
The code changes include the addition of a new module called `trackers` which contains the `Tracker` class for object tracking. The `Tracker` class initializes with a YOLO model and ByteTrack tracker, and provides methods for detecting objects in frames and getting object tracks. The `main.py` file has been updated to use the `Tracker` class to get object tracks from the input video frames. The YOLO model path and stub path are now passed as arguments to the `Tracker` constructor. This commit improves the functionality of the code by adding object tracking capabilities.