feat: Add draw_team_ball_control method to Tracker class
The code changes include the addition of a new method, `draw_team_ball_control`, to the `Tracker` class in the `tracker.py` file. This method draws the team that has control of the ball on a given frame. This enhancement improves the functionality of the `Tracker` class by providing a visual representation of the ball control in the video frames.