feat: Add TeamAssigner class for player-team assignment
The code changes include the addition of a new module called `team_assigner` which contains the `TeamAssigner` class for assigning teams to players. The `TeamAssigner` class initializes with empty dictionaries for team colors and player-team assignments. It provides methods for training a KMeans clustering model on an image, getting the average color of a player, assigning team colors to players in a frame, and getting the team assignment for a player. This commit enhances the functionality of the code by adding the `TeamAssigner` class for player-team assignment.